Add force option to charmcraft init

+++ charmcraft/commands/init.py
@@ -74,13 +74,16 @@ class InitCommand(BaseCommand):
             "--series", default="kubernetes",
             help="The comma-separated list of series this charm will support;"
             " defaults to 'kubernetes'.")
+        parser.add_argument(
+            "-f", "--force", type="store_true",
+            help="Initialize even if the project dir is not empty. Will not overwrite files.")
 
     def run(self, args):
         args.path = args.path.resolve()
         if args.path.exists():
             if not args.path.is_dir():
                 raise CommandError("{} is not a directory".format(args.path))
-            if next(args.path.iterdir(), False):
+            if any(args.path.iterdir()) and not args.force:
                 raise CommandError("{} is not empty".format(args.path))
             logger.debug("Using existing project directory '%s'", args.path)
         else:
@@ -121,6 +124,8 @@ class InitCommand(BaseCommand):
             template_name = template_name[:-3]
             logger.debug("Rendering %s", template_name)
             path = args.path / template_name
+            if path.exists():
+                continue
             path.parent.mkdir(parents=True, exist_ok=True)
             with path.open("wt", encoding="utf8") as fh:
                 out = template.render(context)
